chore: Refactor EventsController and EventsService

+++ src/events/events.controller.ts
@@ -2,6 +2,7 @@ import {
     Body,
     Controller,
     Delete,
+    ForbiddenException,
     Get,
     HttpCode,
     HttpStatus,
@@ -18,26 +19,17 @@ import {
 } from '@nestjs/common';
 import { CreateEventDto } from './input/create-event.dto';
 import { UpdateEventDto } from './input/update-event.dto';
-import { Event } from './event.entity';
-import { Like, MoreThan, Repository } from 'typeorm';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Attendee } from './attendee.entity';
 import { EventsService } from './events.service';
 import { ListEvents } from './input/list.events';
 import { CurrentUser } from 'src/auth/current-user.decorator';
 import { AuthGuardJwt } from 'src/auth/auth-guard.jwt';
+import { User } from 'src/auth/user.entity';
 
 @Controller('/events')
 export class EventsController {
     private readonly logger = new Logger(EventsController.name);
 
-    constructor(
-        @InjectRepository(Event)
-        private readonly repository: Repository<Event>,
-        @InjectRepository(Attendee)
-        private readonly attendeeRepository: Repository<Attendee>,
-        private readonly eventsService: EventsService,
-    ) {}
+    constructor(private readonly eventsService: EventsService) {}
 
     @Get()
     @UsePipes(new ValidationPipe({ transform: true }))
@@ -57,67 +49,10 @@ export class EventsController {
         return events;
     }
 
-    @Get('/practice')
-    async practice() {
-        // find where 使用object為AND，使用array為OR
-        return await this.repository.find({
-            select: ['id', 'when'],
-            where: [
-                {
-                    id: MoreThan(3),
-                    when: MoreThan(new Date('2021-02-12T13:00:00')),
-                },
-                {
-                    description: Like('%meet%'),
-                },
-            ],
-            take: 2, // limit
-            skip: 1, // offset
-            order: {
-                id: 'DESC',
-            },
-        });
-    }
-
-    @Get('/practice2')
-    async practice2() {
-        // return await this.repository.findOne({
-        //     where: {
-        //         id: 1,
-        //     },
-        //     relations: ['attendees'],
-        // });
-
-        const event = await this.repository.findOne({
-            where: { id: 1 },
-            relations: ['attendees'],
-        });
-
-        // const event = new Event();
-        // event.id = 1;
-
-        const attendee = new Attendee();
-        attendee.name = 'Gura';
-        // attendee.event = event;
-
-        event.attendees.push(attendee);
-
-        // await this.attendeeRepository.save(attendee);
-        await this.repository.save(event);
-
-        return event;
-    }
-
     @Get(':id')
     // 如果要傳入其他參數要使用new ParseIntPipe()如果沒有直接使用ParseIntPipe就可以
     // async findOne(@Param('id', new ParseIntPipe()) id) {
     async findOne(@Param('id', ParseIntPipe) id: number) {
-        // const event = this.events.find((event) => event.id === +id);
-
-        // const event = await this.repository.findOne({
-        //     where: { id },
-        // });
-
         const event = await this.eventsService.getEvent(id);
 
         if (!event) {
@@ -134,29 +69,45 @@ export class EventsController {
     }
 
     @Patch(':id')
-    async update(@Param('id') id, @Body() input: UpdateEventDto) {
-        const event = await this.repository.findOne({
-            where: { id },
-        });
+    @UseGuards(AuthGuardJwt)
+    async update(
+        @Param('id') id,
+        @Body() input: UpdateEventDto,
+        @CurrentUser() user: User,
+    ) {
+        const event = await this.eventsService.getEvent(id);
 
         if (!event) {
             throw new NotFoundException();
         }
 
-        return await this.repository.save({
-            ...event,
-            ...input,
-            when: input.when ? new Date(input.when) : event.when,
-        });
+        if (event.organizerId !== user.id) {
+            throw new ForbiddenException(
+                null,
+                `You are not the authorized to update this event`,
+            );
+        }
+
+        return await this.eventsService.updateEvent(event, input);
     }
 
     @Delete(':id')
     @HttpCode(HttpStatus.NO_CONTENT)
-    async remove(@Param('id') id) {
-        const result = await this.eventsService.deleteEvent(id);
+    @UseGuards(AuthGuardJwt)
+    async remove(@Param('id') id, @CurrentUser() user: User) {
+        const event = await this.eventsService.getEvent(id);
 
-        if (result?.affected !== 1) {
+        if (!event) {
             throw new NotFoundException();
         }
+
+        if (event.organizerId !== user.id) {
+            throw new ForbiddenException(
+                null,
+                `You are not the authorized to remove this event`,
+            );
+        }
+
+        await this.eventsService.deleteEvent(id);
     }
 }
